Refactor message components into unified Message module
- Consolidate MessageCard and StreamingMessageItem into src/components/Message/
- Update ChatView imports to use new unified module structure
- Add refactoring documentation for component reorganization
