+++ src/components/Message/index.ts
@@ -1,0 +1,9 @@
+// Message components
+export { default as MessageCard } from "./MessageCard";
+export { default as StreamingMessageItem } from "./StreamingMessageItem";
+
+// Shared components
+export { default as MarkdownRenderer } from "./shared/MarkdownRenderer";
+export { default as ToolCallsSection } from "./shared/ToolCallsSection";
+export { default as ProcessorUpdatesSection } from "./shared/ProcessorUpdatesSection";
+export { default as TypingIndicator } from "./shared/TypingIndicator";
